+++ src/wirestate-react/index.test.ts
@@ -31,6 +31,7 @@ describe("Library exported API from wirestate-react", () => {
     "useOptionalSyncQueryCaller",
     "useQueryCaller",
     "useQueryHandler",
+    "useRootContainer",
     "useScope",
     "useSyncQueryCaller",
   ];